Add NavBar component
A new NavBar component has been created and added to the components folder. This includes a brand title and links for Login and Sign Up with a specific design defined through classes. It improves the structure and readability of the code by separating the navigation bar as a separate component.